requirements.txt
hydra-core==1.3.2
lightning==2.3.2
pandas==2.2.2
pillow==10.4.0
torchvision==0.18.1
tensorboard==2.17.0
numpy==1.26.4